refactor: update code, and enums
