+++ Equia-RandomUserAPI/src/enums/gender-enum.ts
@@ -1,0 +1,5 @@
+export enum Gender {
+  MALE = 'male',
+  FEMALE = 'female',
+  ALL = 'all',
+}
